+++ src/main/kotlin/dev/encore/intellij/runconfig/EncoreRunConfig.kt
@@ -8,12 +8,12 @@ import com.goide.execution.testing.GoTestRunConfiguration
 import com.intellij.execution.configurations.RunnerSettings
 import com.intellij.execution.target.TargetedCommandLineBuilder
 import com.intellij.execution.target.value.TargetValue
-import com.intellij.openapi.project.rootManager
+import dev.encore.intellij.utils.isInEncoreApp
 
 class EncoreRunConfig : GoRunConfigurationExtension() {
     override fun isApplicableFor(configuration: GoRunConfigurationBase<*>): Boolean {
         if (configuration is GoTestRunConfiguration) {
-            return isEncoreApp(configuration)
+            return isInEncoreApp(configuration.defaultModule)
         }
         return false
     }
@@ -23,7 +23,7 @@ class EncoreRunConfig : GoRunConfigurationExtension() {
         runnerSettings: RunnerSettings?
     ): Boolean {
         if (applicableConfiguration is GoTestRunConfiguration) {
-            return isEncoreApp(applicableConfiguration)
+            return isInEncoreApp(applicableConfiguration.defaultModule)
         }
         return false
     }
@@ -50,18 +50,8 @@ class EncoreRunConfig : GoRunConfigurationExtension() {
             }
         }
         if (useEncoreBinary) {
-            cmdLine.exePath = TargetValue.fixed("encore")
+            cmdLine.exePath = TargetValue.fixed("/Users/dom/bin/encoretmp/build_encoretmp")
         }
         super.patchCommandLine(configuration, runnerSettings, cmdLine, runnerId, state, commandLineType)
-    }
-
-    private fun isEncoreApp(configuration: GoRunConfigurationBase<*>): Boolean {
-        for (folder in configuration.getDefaultModule().rootManager.contentRoots) {
-            val appFile = folder.findChild("encore.app")
-            if (appFile != null && appFile.exists()) {
-                return true
-            }
-        }
-        return false
     }
 }